Modify platform test cases
