+++ cypress-tests/cypress-workspace.config.js
@@ -79,7 +79,7 @@ module.exports = defineConfig({
       "cypress/e2e/selfHost/*.cy.js",
       "cypress/e2e/authentication/*.cy.js",
       "cypress/e2e/workspace/*.cy.js",
-      "cypress/e2e/globalDataSources/*.cy.js",
+      "cypress/e2e/ce/**/*.cy.js",
       "cypress/e2e/editor/app-version/version.cy.js"
     ],
     numTestsKeptInMemory: 1,
