+++ cypress-tests/cypress/commands/commands.js
@@ -57,7 +57,7 @@ Cypress.Commands.add("createApp", (appName) => {
   cy.get("body").then(($title) => {
     cy.get(getAppButtonSelector($title)).click();
     cy.clearAndType('[data-cy="app-name-input"]', appName);
-    cy.get('[data-cy="+ Create app"]').click();
+    cy.get('[data-cy="+-create-app"]').click();
   });
   cy.waitForAppLoad();
   cy.skipEditorPopover();
