+++ cypress-tests/cypress/e2e/ce/globalDataSources/globalDataSourcePermissions.cy.js
